mssmt: fix package race condition w/ concurrent callers
In the package atm, we use a global copy of the empty tree shared across
all package instantiations. As is, if concurrent callers attempt to
insert into _independent_ trees, since there's no locking at the empty
tree/node level, they'll eventually run into a race condition wherein
one caller attempts to read the node key and the other attempts to write
it.

In this commit, we fix this issue by forcing the computation of the node
key (the branch/leaf hash) for each level within the empty tree. We do
this bottom to top, so we've already computed the node key for each
sibling, meaning those cached values can be used directly.