tarorpc: create initial package for gRPC proto compilation+definition
In this commit, we make the base package that will be used to compile
our gRPC protos. We include support for REST, and also the JSON/WASM
stub as well. Finally a docker file is added to allows us to easily
build the packages and ensure that the set of dependencies are always
correct.