cmd/tarod: add initial taro daemon binary along with conf+cli config parsing
In this commit, we use the recently added server to make the first cut
of the daemon binary. Compared to how things are in lnd, in this case we
push all the CLI parsing into this binary, and then translate the items
we create into the set of dependencies for the main server.

In the future, we'll pass in the fundamental interfaces the taro server
needs to function (like chain access, etc, etc).