build: add initial sqlc config file, and Makefile helpers and CI pipeline
In this commit, we add the initial scaffolding needed to use sqlc in the
project. This includes the sqlc.yaml file, a script for running the sqlc
docker container to perform code gen, a series of Makefile directives to
generate+migrate from our base SQL, and also updates to the CI pipeline.