tarodb: introduce new package, add SQL schema+queries for macaroon storage
In this commit, we introduce the new tarodb package, which'll house
concrete implementations of the various storage interfaces we'll use for
persistence in the project.

The macaroon storage is a single table, as it only needs to track the
set of root keys and their IDs. We provide sample migrations to
create+delete the set of stored tables.

We nest things under a sqlite package, though this isn't 100% necessary,
as if we use a subset of postressql, then with either some build tags or
type params, we should be able to use both databases within the project.

This commit also includes the generated code from the SQL queries. As we
use the emit_interfaces field, we also get a nice sqlite.Querier
interface we can use for unit tests in place.