tarodb: add persistence wrapper for bakery.RootKeyStore interface
In this commit, we add a persistence wrapper for the bakery.RootKeyStore
interface. This implements the main interface we'll need in order to add
macaroon support to the main daemon.

We also create a helper test method for creating a test sqlite3 instance
for testing. We use the modernc.org sqlite implementation which is 100%
cgo free.