asset: Properly handle empty PrevID (not nil) in asset witness
A recent commit introduced a change to create new assets with a genesis
witness by default, which consists of a non-nil PrevID of all zeros.
`PrevID` contains a `btcec.PublicKey`, which is not a valid point when
its X and Y coordinates are both zero, leading to a failure upon
deserialization.

Now, upon deserializing, we'll inspect the bytes to be read and skip
parsing the `btcec.PublicKey` when it's all zeros.