tarodb: use the golang-migrate library for schema creation
In this commit, we evolve upon the prior approach (where we use the
embedded file system to embed the schema/migrations), and take it on e
step further to use the `golang-migrate` library. Under the hood, the
migration library will create a new table called `schema_migrations`
that it'll use to keep track of the current database version, and if the
migration was partially applied or not.

Usage of this library is also safer than the prior approach, as this'll
ensure everything is run in a single db transaction. It'll also do
things like table locking for the `postgres` backend.