commitment: fix taro commitment key bug related to family keys
In this commit, we fix a bug that would cause invalid Taro commitments
to be generated when creating a new commitment that had a mix of assets
w/ and w/o a family key defined. If an asset doesn't have a family key
defined, then the `TaroCommitmentKey` will return all zero bytes, which
will then cause all the assets that don't have a family key to collide
in the `assetCommitments` map, meaning only the last one will actually
be created.

To fix this issue, we properly set the assetID to the derived genesis
asset ID above. A new test case and assertion has been added that fails
in master without this commit.

We also add a new test case that directly isolates this failure
scenario. If run on master, we'll get this failure:
```
=== RUN   TestTaroCommitmentKeyPopulation
    commitment_test.go:416: commitment has blank taro commitment key!
    commitment_test.go:422:
        	Error Trace:	commitment_test.go:422
        	Error:      	Received unexpected error:
        	            	#2: failed on input commitment.assetDescription{HasFamilyKey:false, IsCollectible:true}
        	Test:       	TestTaroCommitmentKeyPopulation
```