asset+commitment: add new GenesisSigner interface, switch to using key descs everywhere
In this commit, we add a new GenesisSigner interface which allows us to
not have to directly handle a raw private key to sign the assetID when
the key family is used. A version that accepts a raw private key is
added as well to facilitate unit tests.

Along the way, rather than store a raw public key for the main family
key, we start to instead store a key descriptor. This change will allow
us to eventually use the set of wallet and signer interfaces for the lnd
signing backed.

As a result of this change, we need to update the TLV serialization: we
still want to write the raw public key everywhere, but we retain the key
descriptor information if it's set properly.

By using key decs, we ensure that we have all the information needed to
sign with the script keys and also key families associated with assets.