tarogarden: add new taro garden package with fundamental types/interface
In this commit, we introduce a new package: tarogarden. This package
will be used to implement batched asset minting/creation. The primary
interfaces used are the ones very familiar in the lnd/loop/pool
projects.

The most important interface here is the main storage interface, along
with the "driver" scaffolding that will later allow us to write systems
level tests against the set of registered drivers.