tarodb: set up scaffolding for sqlite pragma options, active foreign key constraints
In this commit, we add some basic scaffolding to allow us to specify the
sqlite pragma options, which can affect things like if the WAL mode is
on or not. For now, we just activate foreign key constraints to ensure
that if we have an issue related to foreign keys, then an error will be
returned during routine operations.

The pragma options themselves are accepted as a series of query string,
which are tacked onto the main DSN.