tarodb/sqlite: add initial table schema structure for asset minting+recv
In this commit, we add an initial set of tables and indexes to implement
storage for assets, transactions, managed UTXOs, and also minting
batches.

Batches are used to gather several assets into one, with seedlings
(assets not yet attached to a transaction) referencing them. We use a
central table to store the set of internal keys we use, which are
identifier by an unrolled keychain.KeyDescriptor like structure. The
internal key for a batch will eventually be used as the actual taproot
internal key for the transaction that anchors the batch and creates the
set of assets.

Assets are stored over a few tables with the genesis point, the base
asset fields (type name, etc), the unique asset information (value,
locktime, etc) and the input witnesses each living in separate tables
with appropriate references.

All assets are anchored by a managed UTXO. A managed UTXO is a UTXO that
was initial funded/borrowed from an external walet. This output is then
imported into the core wallet to ensure that it can track the satoshis
amount which is used to ceata a non-dust output that anchors the Taro
asset. This managed UTXO prereset the _latest_ state of the asset and
uses a series of references to identified which UTXO it bleongs to along
with where that is in the chain itself.

This design allows for an eventual future where either the external
wallet is aware of the taro assets (via the tapscript sibling hash), or
the tarod system assumes more ownership of regular on chain UTXOs.