tarodb: activate WAL mode for sqlite by default
In this commit, we activate the WAL mode which ensures that readers
don't block writers and the other way around. Otherwise w/o this we'll
get the SQLITE_BUSY error at times, meaning we need to implement manual
retries. We'll likely need some logic there for write transactions as
well, but this lets us get pretty far.