commitment: add test case for error case of no external splits
One other case we need to address is fully spending an input Taro UTXO using the on-chain address flow: in this case the sender no longer has an asset to use to commit to the new split set. They may need to retain a dummy zero value asset split just for that transition, then can drop it later.