proof: move to self-describing values for the tapscript sibling pre-images
In this commit, we move to self-describing values for the tapscript
sibling pre-images. This change lets us get rid of the current logic
where we derive all possible combinations then check that one of them
matches. Instead we add a new struct that contains a type of leaf vs
branch. All logic is then updated to use this new struct in place.

One needed follow up here is additional tests, particularly for the BIP
86 case.