mssmt: add a new compacted version of the full SMT
This commit introduces the CompactedTree which is a FullTree equivalent
MS-SMT implementation with the added benefit of reduced storage access.
The key intuition is that the tree attempts to omit branches that would
only hash up against the leaf and move the leaf from the last level to
the topmost node of a subtree (in most cases).