proof: add additional information to the AssetSnapshot verify result
This additional information will be useful for importing an asset as
described in the final state transition of a proof file.

There's a lingering TODO here as we don't have the block height
information here, but it can be obtained from the p2p network or a
backing full node. This should be done when we go to also verify the
full header information based on our on-disk state.