chain_bridge+tarogarden: fix chain confirmation
We need to add error handlilng to the confirmation notification
listener. This can be done by using and returning the errChan. This was
discovered when we cancelled the context too early by accident. An error
was logged in lnd but never delivered to us so we blocked here
indefinitely.

We now listen for an incoming confirmation event or error and only then
deliver the received event back to the state machine.