proof: add new Verifier interface to abstract away from proof decode+verify
In this commit we add a new interface to abstract away from the proof
decode+verify process. This'll allow us to slot in new more efficient
verifiers in the future, and also makes it easier to write unit tests as
well.

We move in the existing verification logic into this file to unify
things under the new file name.