tarodb: use UPSERT to handle duplicate chain_txns, add import test case
In this commit, we handle an edge case in the current import logic: if a
chain transaction already exists in the database, then an error will be
returned. An example of this is multiple assets being imported in the
same batch.

To fix this, we make `InsertChainTx` into an UPSERT to return the
primary key of the transaction we already inserted.

We also add an extra clause in our unit tests that inserts the chain
transaction ahead of time, which will fail without the other changes in
this commit, due to a constraint violation.