tarodb: make the InsertInternalKey UPSERT actually a noop
In this commit, we fix a bug that would cause the primary key of an
existing row to be mutated. An UPSERT was aded for internal keys, so we
can gracefully handle duplicate insertions. However, the UPSERT clause
incorrectly modified the _existing_ primary key with the key that
would've been used.

To fix this, we swap out the raw_key field, which is what would violate
the constraint, making this actually a noop.

We also add a test case to exercise this fix.