tarodb: modify InsertManagedUTXO to be an UPSERT
Upserts are very convenient for DB objects that have a unique
identifier. Because two bitcoin UTXOs can never have the same outpoint,
we can make that the unique identifier. An UTXO also can never exist
without its outpoint being known, so we remove the NOT NULL clause.
Because the UNIQUE keyword already creates an index, we no longer need
the explicit coin_index.