tarodb/smt: add namespace field to existing node table
In this commit, we add a namespace field to the existing node table.
This'll allow us to have multiple independent trees live on disk. An
upcoming use case is maintaining a Universe for each set of assets we've
minted.

We also modify the primary key of the table itself to be (hash_key,
namespace). This'll let us do things like copy over a tree to an
entirely new namespace.