tarodb/smt: add new mssmt_roots table to track tree roots
The roots table assumes that each namespace should only have a single
root. So the namespace is the primary key in this instance.

We then have a dual reference into the main smt nodes table, which
ensures that we can have things always be consistent: if a root is
deleted in the node table, then the reference should be also. This
ensures we have no dangling roots.