tarodb: extract common code into upsertAssetsWithGenesis
This commit extracts the common code for inserting/updating a list of
assets with their genesis information into a common function.

NOTE: This commit is not strictly needed as I realized too late that I
don't need to insert the asset in the custodian because importing the
proof already does that. But since it's a nice refactor and removes a
TODO I decided to keep the commit.