tarodb: reduce duplication in asset transfer row, return asset deltas
In this commit, we modify the asset transfer table to just have a ref to
the managed UTXO. This points to the chain txn, so we can get anything
we need from that.

We also now return the set of spend deltas as well. This isn't needed
yet, but we might as well return all the information we can.