multi: add new struct to house tweaked script key information
In this commit, we add a new struct to house the tweaked script key information. This new struct lets us leave the old scriptkey.pubkey as the main key we always use to encode and also verify signatures against.

The new struct is then embedded within the ScriptKey struct as a pointer, so anytime we're comparing to wire assets, it'll match up properly.