tarodb: move the script key tweak inside the asset itself
In this commit, we move the script key tweak to live inside the asset
rather than within the internal key. This matches the existing managed
utxo struct that contains the tweak (sibling and taro root) inside the
row, which is then used to construct the actual key that the managed
UTXO commits to.

We also do this for the address as well.

In the future, this'll allow us to use the tweak as a foreign key in
another row that contains the actual scripts we're committing to.