tarodb: extend TestImportAssetProof by exporting proof after import
This test fails as is, since it relies on using the actual script key
instead of the raw script key to fetch the proof on disk.

The next commit will resolve this by storing both the tweaked key and
the regular key similar to the way we handle key families on disk.