multi: switch to using a distinct script_keys table for assets
In this commit, we modify things to use a distinct table for script_keys. Moving to this surfaced a bug for which a test case was added in the past commit. This move mirrors the family keys where we store the tweaked key and the raw key, along with all the items we need to re-derive the key.

This move also surfaced an issue in the existing itests which import a proof right at the very end: we only have the tweaked script key, so the import just mirrors the state and doesn't actually let us spend the assets. For the addr send case, we've already created the actual script key, so we can use the tweaked key in the new asset to find the actual ID we need.