proof+tarodb: use SerializedKey for AssetBlob
We shouldn't use a struct as a map key that isn't always comparable. The
x and y coordinates of a public key can be represented in memory in
different ways, even if the actual serialized value might be the same
(see FieldValue.Normalize()).
To make sure we don't introduce very hard to find bugs, we use the
serialized format of public keys in structs that we use as map keys.