tarogarden: add TODO about exclusion proofs and FundPsbt
Currently the FundPsbt returns a TX that has a change output from the
default scope (which is BIP84). And even if it returned a P2TR change
output, the internal key information in the PSBT output for that would
be missing. So we need a change in lnd to fix this.