build: update to sqlite version 1.19.1
We were on a version a few months old.

Main driver of this was I ran into this issue in another branch:
https://gitlab.com/cznic/sqlite/-/issues/77.

Updating fixes it, but then runs into some other related race. The
author of the library asked about the first on the sqlite forums and
that the race also exists in the orginal C codebase, but it isn't an
issue.

The tests fail as is, since there was a bug fix w.r.t how the NULL
constraints and NULL insertion was handleed. It looks like this was
resolved here: https://gitlab.com/cznic/sqlite/-/issues/100 and here
https://gitlab.com/cznic/sqlite/-/issues/98. Seems like the classic case
of `nil` vs `[]byte{}`.