tarodb: restore old version of assetWitnessEqual
The new DeepEqual methods added won't work properly. For example, in the
TestImportAssetProof test, we use the randomly generated asset as the
root asset in a split commitment. We serialize this as a raw asset
within the split commitment proof. However, we have certain fields we
maintain in memory in the struct that are never serialized. An example
of such a field is the raw family key. As a result, these two assets
will never return true from `DeepEqual`. For now, we go back to
asserting against the serialized versions.