tarogarden: fix bug that would cause a nil taro root to be written
The update to sqlite fixes an issue where a nil byte slice would be
stored/interpreted as a `[]byte` instead of a zero sized blob.

As a result, after a restart, we would be passing down a NULL Taro root
to the DB. With the older version this would be written, but the newer
version now properly rejects this as a NULL constraint violation.

We now also cache the script root and return that along side the key.