asset+tarodb: fix DeepEqual of split commitment
Fixes an intermittent test failure introduced by an incorrect version of
the DeepEqual method of split commitments.
For example, in the TestImportAssetProof test, we use the randomly
generated asset as the root asset in a split commitment. We serialize
this as a raw asset within the split commitment proof.
However, we have certain fields we maintain in memory in the struct
that are never serialized. An example of such a field is the raw family
key. As a result, these two assets will never return true from
the old version of `DeepEqual`.
This commit fixes the comparison by also encoding the root asset in the
split commitment before comparing it.