multi: add new proof.Courier interface for async proof distribution
In this commit, we add a new `proof.Courier` interface that's optionally
used as async proof distribution. The current default implementation
uses a hashmail server to send the proofs back and forth between the
sender and the receiver. In the future, this'll be replaced with
streaming/long-polling connections to a desginated multi-verse.

Fixes #95