multi: refactor proof file format
With this commit we change the proof file's format to a proof chain.
Every proof is hashed with its previous proof's hash (the first
proof will have a zero prev hash) to get to the proof's rolling
checksum. That way a file can be read serially, proof-by-proof, without
needing to load the full file content into memory.
This also allows us to append a proof to an existing file easily, once
we have a seekable proof reader that can read just the last proof of a
file.