tarodb+tarofreighter: properly handle case when send can't succeed
In this commit, we add a new error that'll be returned from the coin
selection interface if the set of constraints specified can't be met.
This error will now properly be propagated up to the RPC interface and
thru the command line instead of crashing due to a OOB access (returned
slice has no asset inputs).