tarofreighter: redo fee accounting for PSBT send pkt after adding input
In this commit, we fix an existing TODO to ensure that the final packet
we have has enough fees at the target fee level. Before this commit,
when we added our own input, we would cause the final fee to undershoot
the desired fee rate target. In order to fix this, we'll now compute the
total weight and then required fee (at the fee rate) to meet our target.
We'll then add this delta (`target-current`) to make sure we we're not
undershooting in fees.