tarofreighter+tarodb: record and use height hint
This fixes an issue that would cause us to miss a tx confirmation in the
case where a block came in between we published the transaction and
registered for block confirmations, or we resumed a pkg after shutdown.

We fix this by recording block height before we publish the transaction,
such that we'll use this whenever we later register for confirmations.