tarodb: move key family filter into the WHERE clause for QueryAssets
In this commit, we fix a bug that would cause _all_ other assets
matching some of the sub-queries to be returned along side an asset with
a given key family. Before this commit, we'd use a LEFT JOIN at the top
for the key family information, which meant all other assets would be
returned but with nil information if it didn't match the join.

Instead, we move the filtering down to the WHERE clause which ensures
that if we're trying to filter by key family, then we get _only_ those
assets which match.