tarogarden+tarodb: ensure asset seedling names are unique
This commit adds a unique constraint to the SQL field `asset_seedlings.asset_name`.

A new error type is added to tarogarden to help handle seedling name collisions.

An integration test ensures that an error is returned when attempting to mint an asset whose name collides with an existing minted asset.

This commit also adds functionality for generalising SQL errors across database backends.