internal+vm: remove invalid check, add script tree spend tests
With this commit we remove an invalid check that was added early on that
isn't actually used and breaks spending an asset with a control block in
the witness (since the serialized control block with a one element
inclusion proof is exactly 65 bytes by coincidence).

To make sure signatures and sighash flags are working as expected, we
extend our state transition unit tests with tests that use a script tree
and inclusion proofs in the control block.