proof: refactor tapscript sibling hashing, fix decoding err
This commit extracts the hashing of a tapscript sibling into a method on
the TapscriptSibling struct for easy re-use.
We also add two helper functions for creating a sibling from either a
tap leaf or a tap branch.
Finally, we increase our test coverage of the touched code and fix a
leaf decoding issue that was uncovered by the tests.