commitment: remove outdated comment in commitment
With this commit we remove a comment in the commitment package that does
not really make sense anymore. The tapscript sibling is a specific field
in the proof file. We only really care about the sibling preimage not
the full merkle proof of the tapscript tree (at least from the point of
view of the Taro protocol).