asset+proof: move asset encoder to avoid circular dep
We need to use the asset leaf encoder and decoder in the taropsbt
package. This isn't possible as long as they are in the proof package,
because that would create a circular dependency. So we move the
encoder/decoder to the asset package where they belong to in the first
place.