asset: remove existing varbytes Meta field in place of MetaHash
In this commit, we remove the existing (effectively unbounded) meta
field and replace that instead with a MetaHash which is a fixed sized
field. This ensures that if the meta is very large (say 100 KB), then
the proof size isn't impacted as a result. With this change, the meta
pre-image would then _optionally_ be revealed within the proof file
itself. As a result, the meta only impacts the size of the very first
state transition.