proof: add optional MetaReveal TLV type to proof format
In this commit, we add an optional MetaReveal TLV type to the proof
format. This allows a genesis asset to reveal what was committed to
within the MetaHash in the original asset. This verification would only
be done for genesis assets, as this field cannot be present for any
other asset state transition.