tarodb: add new assets_meta table to store meta reveal + hash info
In this commit, we add a layer of indirection from the asset genesis
(and seedling) to the meta data for an asset. This should make queries
that have a very large meta more efficient, as we no longer need to
fetch the entire meta each time we need to do a query.

This only updates the main table definitions and view. In a later
commit, we'll also update the full set of queries.