cmd/lncli: update minting call with new meta API, add option to read from file
In this commit, we update the CLI commands to account for the changes to
the meta API. Along side specifying the meta blob/bytes, users can now
also specify the type, which only has one actual defined value. We also
add the option to read the meta from a file itself, as very large blobs
can't easily be passed over the CLI itself.