server: stop returning an error that is always nil
The `NewServer` function signature is `func(cfg *Config) (*Server, error)`
but the error is always `nil`. Even in other projects like `loop`, `pool` or
`lnd` we did not find a case were an error needed to be returned.

We can delete it from the signature from now and add it in the future
(or in a wrapped function) if needed.