universe: add new fundamental root universe interfaces
In this commit, we add the set of fundamental interfaces for the
root/base universe. The universe contains all the minting outpoints for
a given asset ID or group key. The leaves of the tree are keyed by a
hash of the minting outpoint (UTXO at the time of minting) and the
script key for the asset within the leaf.

Issuance is registered using a normal genesis proof along side a key.
After issuance, the leaf is inserted into the tree, with a new inclusion
proof returned.