tarodb: add basic tests for persistent base universe
Along the way, we needed to modify the `namespace_root` value to be a
deferred constraint. This allows us to delete an smt root, then re add a
pointer w/o triggering a foreign key constraint violation.

The unique set of values in a universe leaf (as there can be many leaves
for a given universe), os now the minting point and script key within
that.