multi: only use subset of addr for proof delivery
In order to make it easier to transfer proofs for non address sends, we
declare a new Recipient struct which only houses the subset of address
fields that are actually required for delivering a proof.
This allows us to be more generic in the porter, even if we don't use
the hashmail delivery for interactive sends.