multi: refactor transfers to show inputs and outputs
This commit is the main meat and potatoes of the whole database
refactor. Unfortunately it's quite large as it changes structs used in
many areas.

Since this commit also changes the model we use to identify assets and
transfers, we want to document it here. The following NEW rules are
introduced with the code in this commit:

1. Each transfer represents a state update of a single asset ID. There
   can be multiple inputs (=coins of the same asset) and outputs
   (=splits) but only related to the same ID. The exception here are
   passive assets, which are handled in a separate struct/table.
   TODO: Find out how this needs to change to support sending of
   multiple coins belonging to the same asset group but with different
   IDs.
2. Each script key is identified to either be local (it can be derived
   by the lnd node connected to the local daemon) or not.
3. The inputs to a transfer remain in the assets table but are set to
   being spent. That means assets are no longer updated in-place, except
   for passive assets.
4. Transfer outputs are turned into asset entries in the database once
   the transfer confirms. BUT only if their script key is identified as
   being local. That means that transfers with a zero amount tombstone
   that use the NUMS point as the script key don't appear in the assets
   table anymore. The same is true for assets that are sent out of the
   daemon to a remote recipient.

The above rules should make it no longer necessary to filter out split
send tombstones from coin selection, since they no longer appear as
assets.