address: add raw AssetID to address.Taro struct
In this commit, we add the raw AssetID to the address.Taro struct. This
is a prep to remove the genesis bootstrap, as with the Universe flow,
one needs to have verified an asset before you can receive it.

We also modify the `TaroCommitment` to exit if the asset genesis isn't
set. We do this to catch bugs early that would generate the wrong
template asset as the asset genesis wasn't fully known.