make: add unit-debug and unit-trace targets
We add two new ways to run unit tests:

unit-debug: Add -v to the "go test" call, causing it to show tests
executed as well as the logs generated by calls to "t.Logf()".

unit-trace: Same as above, but we also enable the package level loggers
that print to stdout with the trace level set.