multi: address TODOs around commitment pruning
This commit updates the interactive full value send integration test to
also attempt to send the passive asset at the end. This demonstrates
that the pruning of an empty asset tree in the Taro tree was missing
before, so this commit fixes it. At the same time we also fix another
TODO around removing asset commitments, as we now have a way to detect
an empty tree.