tarodb: update seedling to add group anchor
In this commit, we extend the seedling type to add an optional field,
a group anchor. A group anchor is another seedling that will be used
as the first asset in the group, and therefore the initial tweak for
deriving the group key. Seedlings that specify a group anchor will be
minted into the same asset group as the group anchor and other
seedlings that use the same anchor.