tarodb: add FetchSeedlingID query
In this commit, we add queries to support fetching the database ID of
a seedling, and vice-versa. This is needed to represent the use of a
group anchor without duplicating the anchor name for every seedling
that specifies an anchor. Seedling name uniqueness is enforced in
other areas of the planter, so we can safely map names to IDs.