tarodb/sqlc: add explicit length check for keys in universe tree
In this commit, we commit to using the schnorr serialization in the
Universe tree. We use it for the top level key as this is what's used to
derive the taro commitment key. We also use it for the script key as
well since we're able to save a byte, and the actual proof has the full
parity information.