tarodb: use "_txlock=immediate for sqlite transactions
In this commit, we switch to using `"_txlock=immediate` as only with
this will the busy timeout actually be used. With the latest update to
sqlc, it'll only use the immediate mode if a transaction is marked as
write.

We also modify the Get call for the key store to use a transaction.