tarodb: set values for max and idle conns for sqlite
In this commit, we set values for max and idle conns for sqlite. This
ensures that with the new transactional mode, we don't need to always
block for transaction to be executed.

We also set the synchronous value to full (fsync after writing the WAL),
and also set fullfsync which ensures that transactions on mac are
properly durable.