tarodb: remove instance of non-txn based db access
In this commit, we fix an existing bug wherein we'd use the raw database
to issue queries, which ended up causing other transactions to fail
later with the new transactional model. Before we'd start to do a query
(to prep for a test or w/e), but those changes would actually still be
uncommitted. As a result, another transaction that needed to modify that
value wasn't allowed to continue.

One instance of this was related to the macaroon calls, while the other
instance we within some unit tests.