taropsbt: properly encode BIP32 derivation paths
BIP174 defines the encoding of the BIP32 derivation paths (including the
Taproot specific ones) as containing the public key within the key part
of the PSBT field. That allows a PSBT to contain multiple entries for
the derivation paths, which we are going to allow for the virtual PSBTs
in a follow-up commit as well.