multi: remove tapscript sibling from addr
The tapscript sibling functionality was never fully implemented in
combination with an address. And we would need the tapscript _preimage_
in the address in order to be able to correctly create the proper
inclusion proofs when sending to an address. That will require a BIP
update so we remove the incomplete functionality from the code for now.