multi: clarify tapscript preimage, rename taro_root to merkle_root
In order to create the proper inclusion proofs, we always need the
preimage of a tapscript sibling. So we don't just store the sibling's
hash but the TLV encoded preimage.
To make it clear what root is shown in the ListAssets call, we also
rename the taro_root to merkle_root since if there is a tapscript
sibling, then the root shown is actually the merkle tree root hash and
not the Taro root commitment hash.