commitment+proof: add Tapscript sibling and helpers for it
With this commit we add the Tapscript sibling preimage struct to the
asset snapshot and the base proof params, so we can create proper
inclusion and exclusion proofs for outputs with a Tapscript sibling.

Because we need to use an encoded version of the Tapscript sibling in
the PSBT and database, we also add helper functions for safely
encoding/decoding the optional sibling.