commitment: allow multiple splits in same anchor output
To allow multiple internal splits (send to self) in a single anchor
output, we need to remove some of the restrictions added early on.
This also removes an old TODO where we can now answer: "Yes, there
are reasons to allow internal splits to be mapped to the same external
output".