mssmt: overwrite Copy of CompactedLeafNode
This fixes a problem that caused a copy of a CompactedLeafNode to be
different from the original. Because we didn't overwrite the Copy method
before, the one from the embedded LeafNode was used, which resulted in
an inaccurate/incorrect copy.
By overwriting the Copy method we make sure that the resulting copy is
completely identical to the original.