server: add `ValidateMacaroon` to server
When tarod is executed as a standalone process the interceptorChain is
in charge of checking the macaroon validity when an endpoint is hit by a
client. However, for we want the server to be able to check the validity
of a macaroon by itself if needed. This is useful when the server is
running as part of another pocess (ex: LiT).