multi: add Taro root alongside Merkle root
Because the Merkle root is not equal to the Taro root when we have a
tapscript sibling, it is useful to know both values.
Especially if we intend to spend the BTC anchor output using the
tapscript path, we need to know the Taro root as that needs to be
included in the inclusion proof of the control block within the witness.