universe: log universe syncs + proof insertion
In this commit, we integrate the new Telemetry interface to log events
for new universe sync and proof insertion events. We log all proof
insertion, but actually log each leaf fetched as a sync event.
Alternatively, we could log when either the set of leaves or keys are
fetched, but logging for leaves instead will give us a more accurate
view on the set of sync stats.