universe: add new FederationEnvoy sub-system
In this commit, we add the FederationEnvoy sub-system that's responsible
for syncing proofs and also pushing proofs between a federation of
universe servers.

Each tick, we try to sync with the federation in a parallel manner.

The new sub-system also implements the universe.Registrar interface.
This allows the minting service to push new proofs to the envoy,
which'll then add it to the local registry before pushing it out to the
rest of the federation.