tapcfg: increase the receiver ack timeout, and number of attempts
We want the hashmail server to be as async as possible, so we increase
the number of attempts to 2k, this is around 2 weeks to deliver a proof
before giving up (groutine exits, but can restart to kick again, and in
the future can do so via RPC).

In addition, we increase the receiver timeout to 6 hours. Otherwise, as
is we'll start the back off (prior was only 4 attempts) after only 10
seconds of waiting. If the receiver isn't at that point, then we'll
quickly bleed out attempts.