multi: allow batch key to be specified as hex
This is a follow-up to #286 which allowed developers to specify any raw
bytes as a hex string in the REST API (instead of the library default of
base64).
The only area where that change didn't apply is the URI part of REST
calls. For those we need to use a string field to support hex encoding,
the same way we already did for the universe REST API.