tapgarden: use private atomic uint for batch state
In this commit, we change the MintingBatch.BatchState variable from a
public uint32 to a private atomic.uint32, to prevent race conditions
when cancelling a batch. We also add methods to read and write this
private batch state, and new logic to check that plain uint32s are valid
BatchStates.