tapdb: use the serializable isolation by default for postgres
In this commit, we start to use the serializable snapshot isolation by
default for postgres. For sqlite, this is a noop as only a single writer
is allowed at any time. With this mode, we should guard against all
types of anomalies, at the cost of extra locking and potential perf
hits.

One thing we'll need to test out is if the DB driver will automatically
do retries or not if a transaction cannot be serialized.

This is an attempt at fixing #333 at a global level.